+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Transition.hs
@@ -1,18 +1,23 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Alonzo.Transition (
   TransitionConfig (..),
+  alonzoInjectCostModels,
 ) where
 
+import Cardano.Ledger.Alonzo.Core (AlonzoEraPParams, ppCostModelsL)
 import Cardano.Ledger.Alonzo.Era
-import Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis)
+import Cardano.Ledger.Alonzo.Genesis
 import Cardano.Ledger.Alonzo.Translation ()
 import Cardano.Ledger.Mary
 import Cardano.Ledger.Mary.Transition (TransitionConfig (MaryTransitionConfig))
+import Cardano.Ledger.Plutus.CostModels (CostModels, updateCostModels)
+import Cardano.Ledger.Shelley.LedgerState
 import Cardano.Ledger.Shelley.Transition
 import GHC.Generics
 import Lens.Micro
@@ -27,7 +32,8 @@ instance EraTransition AlonzoEra where
 
   mkTransitionConfig = AlonzoTransitionConfig
 
-  injectIntoTestState = shelleyRegisterInitialFundsThenStaking
+  injectIntoTestState cfg =
+    shelleyRegisterInitialFundsThenStaking cfg . alonzoInjectCostModels cfg
 
   tcPreviousEraConfigL =
     lens atcMaryTransitionConfig (\atc pc -> atc {atcMaryTransitionConfig = pc})
@@ -36,3 +42,20 @@ instance EraTransition AlonzoEra where
     lens atcAlonzoGenesis (\atc ag -> atc {atcAlonzoGenesis = ag})
 
 instance NoThunks (TransitionConfig AlonzoEra)
+
+alonzoInjectCostModels ::
+  (EraTransition era, AlonzoEraPParams era) =>
+  TransitionConfig AlonzoEra -> NewEpochState era -> NewEpochState era
+alonzoInjectCostModels cfg =
+  case agExtraConfig $ cfg ^. tcTranslationContextL of
+    Nothing -> id
+    Just aec -> overrideCostModels (aecCostModels aec)
+
+overrideCostModels ::
+  (EraTransition era, AlonzoEraPParams era) =>
+  Maybe CostModels ->
+  NewEpochState era ->
+  NewEpochState era
+overrideCostModels = \case
+  Nothing -> id
+  Just cms -> nesEsL . curPParamsEpochStateL . ppCostModelsL %~ updateCostModels cms